rag-service/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-ollama
chromadb
python-dotenv
pypdf
